test(pr-workflow): add F3 anti-templating rejection-path coverage, split lifecycle test file
A Stage B reviewer independently re-raised a gap: F3's anti-templating
distinctness check (workflowArtifactHasContractMarker's multi-lane evidence
comparison) had zero committed regression coverage for its rejection path —
only positive-path (genuinely distinct evidence) coverage existed anywhere
in the suite. The mechanism itself was already correct, but per this repo's
own permanent directive ("untested branches count as unwired code and are
blockers") and the precedent this same PR set for F13/F14/F16/F17 (identical
"real mechanism, no regression test" gaps that were closed rather than
deferred), added a dedicated test: a tier-S consolidated lane owning three
dimensions, persisted with persistBatch's own default text generator (which
produces identical [CLEAN] wording for every owned family beyond the first —
a real copy-paste-relabel case, not contrived), asserting the base-coverage
gate correctly rejects it; then a distinct-text retry asserting it settles.
Verified the test fails when the distinctness check is disabled.

This pushed tests/unit/hooks/pr-workflow-gate-lifecycle.test.ts over the
FR-006 500-line cap (512 lines against a 241-line base — a ratchet
violation). Split it by functional area: general lifecycle (activation,
caching, head-binding, dispatch-gating) stays in the original file; base-
dimension coverage and depth-tier consolidation/persistence (the bulk of
the file's growth across this feedback round) moves to a new
pr-workflow-gate-base-coverage.test.ts. Verified both files pass in
isolation and together with all other consumers of the shared
pr-workflow-gate.test-fixtures.ts module.
